Sketch API for using custom HTML
Following the discussion at https://discourse.elm-lang.org/t/introducing-elm-editor-a-web-based-ide-for-elm-programs/7824/4?u=viir

The idea is to have the development tool interpret a declaration as an entry point/root if the declaration name is `htmlMain`.
The editor could take the value out of this declaration and forward it to the HTML preview.

I have not yet tested if the formatting/escaping in the script tag is complete.

This example (re)uses existing APIs for Fullstack App development to integrate Elm make and read from source files.
Usage of the `CompilationInterface.ElmMake` interface is documented at https://github.com/elm-fullstack/elm-fullstack/blob/78243a5db8349186698d3aecdd4ac7c83496a0e0/guide/how-to-configure-and-deploy-an-elm-fullstack-app.md#compilationinterfaceelmmake-elm-module

This commit only illustrates an idea of how an API could look. Elm Editor does not yet support this kind of entry point.